Fix issue #289
Prevent user from jailbreak. The first time we detect that the device is unsecure (jailbroken or dev mode), we will warn the user with a popup. Next time he launches the application, the user is going to see a snackbar.
Creation of a security manager to have the capability to know the device security (jailbreak or dev mode right now) from every where in the app